pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orius"
version = "0.1.0"
description = "ORIUS: Autonomous Energy Intelligence & Optimization Platform"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Pratik Niroula" }]
license = { text = "MIT" }
keywords = [
  "grid",
  "energy",
  "forecasting",
  "optimization",
  "dc3s",
  "iot",
  "conformal-prediction",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
urls = { Homepage = "https://github.com/Pratikn03/Capstone_Projects", Repository = "https://github.com/Pratikn03/Capstone_Projects", Documentation = "https://github.com/Pratikn03/Capstone_Projects/tree/main/docs" }
dependencies = [
  "altair==5.5.0",
  "annotated-doc==0.0.4",
  "annotated-types==0.7.0",
  "anyio==4.12.1",
  "appnope==0.1.4; sys_platform == 'darwin'",
  "asttokens==3.0.1",
  "attrs==25.4.0",
  "blinker==1.9.0",
  "cachetools==6.2.6",
  "certifi==2026.1.4",
  "charset-normalizer==3.4.4",
  "click==8.1.8",
  "comm==0.2.3",
  "contourpy==1.3.0",
  "cycler==0.12.1",
  "debugpy==1.8.20",
  "decorator==5.2.1",
  "duckdb==1.4.4",
  "exceptiongroup==1.3.1",
  "executing==2.2.1",
  "fastapi==0.128.0",
  "fastjsonschema==2.21.2",
  "filelock==3.19.1",
  "fonttools==4.60.2",
  "fsspec==2025.10.0",
  "gitdb==4.0.12",
  "gitpython==3.1.46",
  "h11==0.16.0",
  "httpx==0.28.1",
  "huggingface-hub==0.36.0",
  "httptools==0.7.1",
  "highspy==1.8.1",
  "hypothesis==6.151.11",
  "idna==3.11",
  "importlib-metadata==8.7.1",
  "importlib-resources==6.5.2",
  "ipykernel==6.31.0",
  "ipython==8.18.1",
  "jedi==0.19.2",
  "jinja2==3.1.6",
  "joblib==1.5.3",
  "jsonschema==4.25.1",
  "jsonschema-specifications==2025.9.1",
  "jupyter-client==8.6.3",
  "jupyter-core==5.8.1",
  "kaggle==1.7.4.5",
  "kiwisolver==1.4.7",
  "lightgbm==4.6.0",
  "markupsafe==3.0.3",
  "matplotlib==3.9.4",
  "matplotlib-inline==0.2.1",
  "mpmath==1.3.0",
  "mypy==1.15.0",
  "mypy-extensions==1.1.0",
  "narwhals==2.15.0",
  "nbformat==5.10.4",
  "nest-asyncio==1.6.0",
  "networkx==3.2.1",
  "ngboost==0.5.8",
  "numpy==2.0.2",
  "openpyxl==3.1.5",
  "packaging==25.0",
  "pandas==2.3.3",
  "parso==0.8.5",
  "pathspec==0.12.1",
  "patsy==1.0.2",
  "pexpect==4.9.0",
  "pillow==11.3.0",
  "platformdirs==4.4.0",
  "plotly==6.5.2",
  "prompt-toolkit==3.0.52",
  "protobuf==6.33.5",
  "psutil==7.2.2",
  "ptyprocess==0.7.0",
  "pure-eval==0.2.3",
  "pyarrow==21.0.0",
  "prometheus_client==0.25.0",
  "pydantic==2.12.5",
  "pydantic-core==2.41.5",
  "pydeck==0.9.1",
  "pygments==2.19.2",
  "pyparsing==3.3.2",
  "pyomo==6.8.2",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.2.1",
  "pytz==2025.2",
  "pyyaml==6.0.3",
  "pyzmq==27.1.0",
  "referencing==0.36.2",
  "requests==2.32.5",
  "rpds-py==0.27.1",
  "ruff==0.9.10",
  "scikit-learn==1.6.1",
  "scipy==1.13.1",
  "six==1.17.0",
  "smmap==5.0.2",
  "stack-data==0.6.3",
  "starlette==0.49.3",
  "statsmodels==0.14.6",
  "sympy==1.14.0",
  "tenacity==9.1.2",
  "threadpoolctl==3.6.0",
  "toml==0.10.2",
  "torch==2.8.0",
  "tornado==6.5.4",
  "traitlets==5.14.3",
  "typing-extensions==4.15.0",
  "typing-inspection==0.4.2",
  "tzdata==2025.3",
  "urllib3==2.6.3",
  "uvicorn==0.39.0",
  "uvloop==0.22.1",
  "watchfiles==1.1.1",
  "wcwidth==0.5.2",
  "websockets==15.0.1",
  "wfdb==4.3.0",
  "xgboost==2.1.4",
  "zipp==3.23.0",
]

[project.optional-dependencies]
dev = [
  "mypy==1.15.0",
  "mypy-extensions==1.1.0",
  "pathspec==0.12.1",
  "ruff==0.9.10",
]
baselines = [
  "prophet==1.1.5",
  "darts==0.30.0",
  "flaml[automl]==2.3.4",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
